Improved blitter emulation from Tob?:
This new code try to emulate the internal blitter processes as described
in the official documentation.
It does not perform any operation if it's not necessary, so it doesn't
need a cycle table, the timings are accurate because the logic is
accurate.
The registers are kept up to date, the words-per-line never reach zero,
and the source and destination registers are incremented as described in
the docs.
When started with 0 in either words-per-lines or lines-per-block, the
value stored in these registers is set to 65536.