install hconsole under /usr/share + add new example
example.py needs hconsole.py to be in same directory
(or in Python module path) to be able to import it.

Hconsole also isn't that useful in itself, it's more of something
on top of which one can build things (or test Hatari remote API),
so it doesn't need to be installed to binary directory. The Hatari
data directory is enough.