update cmdline.c
If there is only one arg and it ends in ".stc", load a config file. Else load disk-a.

Created a new extension "stc" instead of the default "cfg" to enable unique file association.

Amiga and Atari ST playlists work better if they are list of config files instead of disks. UAE was already set up for this. Hatari was not.